Move email jobs to jobs/emails
