--- app/jobs/send_application_submission_email_job.rb
@@ -1,16 +1,0 @@
-class SendApplicationSubmissionEmailJob < ApplicationJob
-  include CourseHelper
-
-  queue_as :default
-
-  def perform(application:)
-    GenericMailer.with(
-      to: application.user.email,
-      full_name: application.user.full_name,
-      provider_name: application.lead_provider.name,
-      course_name: localise_sentence_embedded_course_name(application.course),
-      amount: application.raw_application_data["funding_amount"],
-      ecf_id: application.ecf_id,
-    ).application_submitted.deliver_now
-  end
-end
